third_party/blink/web_tests/platform/generic/http/tests/dom/mutationobserver-parserappend-childlist-expected.txt

CONSOLE MESSAGE: PASS: childList mutation event fired from HTML parser change.
Test result is dumped on console.log. foo bar
